Make MyPermissivePolicy even more permissive
Recently I copied this permissive code thinking that it was completely wildcarded, a nice way to grab a good template for the auth component and scope later when I was done poking around with it. I spent 2 hours trying to figure out what was wrong with it when my client device would give the message "Connection failed with exception AWS_ERROR_MQTT_PROTOCOL_ERROR: Protocol error occurred." I couldn't figure out why I was having a protocol error when I tried out both of Moquette (MQTT v3.1.1) and EMQX (MQTT v5.0) surely this mqtt script that I had just ran successfully a couple months ago wasn't broken.

Lo and behold it was this script causing problems the whole time. I glanced at the greengrass logs at some point but the Device isn't authorized to connect I either didn't see over the wall of errors, or I was preoccupied still perplexed at the MQTT protocol error.

I think for a demo policy that is supposed to be the most permissive policy, we should make it the MOST permissive, I surely can't be the only one this has happened to.